% modification on git from copied files
\name{ratThymus}
\docType{data}
\alias{ratThymus}

\title{Data from female rat thymus tissue taken from the Rat BodyMap
  project (Yu et al, 2014).}

\description{This data set is a matrix ('mobData') of raw count data acquired
  for 17230 genes from 16 samples of female rat thymus tissue. The
  tissues are extracted from four different age groups (2, 6, 21 and 104
  week) with four replicates at each age. Gene annotation is given in
  the rownames of the matrix.
}

\value{A matrix}
\usage{ratThymus}
\format{A matrix of RNA-Seq counts in which each of the sixteen columns
  represents a sample, and each row a gene locus.}
\source{Illumina sequencing.}
\seealso{\code{\link{cD.ratThymus}}}

\references{Yu Y. et al. A rat RNA-Seq transcriptomic BodyMap across 11 organs and 4
  developmental stages. Nature Communications (2014)}

\keyword{datasets}
    
